requirements.txt
appdirs==1.4.3
asgiref==3.6.0
Django==4.2.1
django-appconf==1.0.5
django-cleanup==7.0.0
django-crispy-forms==2.0
django-filter==23.2
django-imagekit==4.1.0
djangorestframework==3.14.0
djangorestframework-jsonp==1.0.2
importlib-metadata==6.6.0
Markdown==3.4.3
pilkit==2.0
Pillow==9.5.0
pytz==2023.3
six==1.10.0
sqlparse==0.4.4
zipp==3.15.0
